replace the logger
