+++ framework/generator/generator.go
@@ -26,7 +26,7 @@ func Generate(state *State) ([]byte, error) {
 	return generator.Generate(state)
 }
 
-func New(bfs *budfs.FileSystem, flag *framework.Flag, injector *di.Injector, log log.Interface, module *gomod.Module, parser *parser.Parser) *Generator {
+func New(bfs *budfs.FileSystem, flag *framework.Flag, injector *di.Injector, log log.Log, module *gomod.Module, parser *parser.Parser) *Generator {
 	return &Generator{bfs, flag, injector, log, module, parser, nil}
 }
 
@@ -34,7 +34,7 @@ type Generator struct {
 	bfs      *budfs.FileSystem
 	flag     *framework.Flag
 	injector *di.Injector
-	log      log.Interface
+	log      log.Log
 	module   *gomod.Module
 	parser   *parser.Parser
