+++ framework/generator/generator.gotext
@@ -27,20 +27,9 @@ func run(ctx context.Context) error {
 	return cli.Parse(ctx, os.Args[1:])
 }
 
-func logger() (log.Interface, error) {
-	{{/* TODO: configurable log level */}}
-	handler, err := filter.Load(console.New(os.Stderr), "info")
-	if err != nil {
-		return nil, err
-	}
-	return log.New(handler), nil
-}
-
 func generate(ctx context.Context) error {
-	log, err := logger()
-	if err != nil {
-		return err
-	}
+	{{/* TODO: configurable logger */}}
+	log := log.New(log.InfoLevel, console.New(os.Stderr))
 	module, err := gomod.Find(".")
 	if err != nil {
 		return err
@@ -48,7 +37,7 @@ func generate(ctx context.Context) error {
 	generators, err := {{ $.Provider.Name }}(
 		{{- if $.Provider.Variable "context.Context" }}ctx,{{ end }}
 		{{- if $.Provider.Variable "github.com/livebud/bud/package/gomod.*Module" }}module,{{ end }}
-		{{- if $.Provider.Variable "github.com/livebud/bud/package/log.Interface" }}log,{{ end }}
+		{{- if $.Provider.Variable "github.com/livebud/bud/package/log.Log" }}log,{{ end }}
 	)
 	if err != nil {
 		return err
