+++ package/hot/client.go
@@ -19,12 +19,12 @@ import (
 // from the following minimal eventsource stream:
 // - https://github.com/neelance/eventsource/blob/master/client/client.go
 // Thanks Richard!
-func Dial(log log.Interface, url string) (*Stream, error) {
+func Dial(log log.Log, url string) (*Stream, error) {
 	return DialWith(http.DefaultClient, log, url)
 }
 
 // DialWith creates a server-sent event (SSE) stream with a custom HTTP client.
-func DialWith(client *http.Client, log log.Interface, url string) (*Stream, error) {
+func DialWith(client *http.Client, log log.Log, url string) (*Stream, error) {
 	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
 		return nil, err
@@ -47,7 +47,7 @@ func DialWith(client *http.Client, log log.Interface, url string) (*Stream, erro
 }
 
 type Stream struct {
-	log     log.Interface
+	log     log.Log
 	res     *http.Response
 	eventCh chan *Event
 	errorCh chan error
