+++ package/log/console/console_test.go
@@ -1,9 +1,12 @@
 package console_test
 
 import (
+	"errors"
+	"os"
 	"testing"
 
-	"github.com/livebud/bud/package/log2/console"
+	log "github.com/livebud/bud/package/log"
+	"github.com/livebud/bud/package/log/console"
 )
 
 func TestConsole(t *testing.T) {
@@ -12,4 +15,7 @@ func TestConsole(t *testing.T) {
 	console.Field("file", "console_test.go").Field("another", "cool story").Notice("hello %s", "mars")
 	console.Field("file", "console_test.go").Field("another", "cool story").Warn("hello %s", "mars")
 	console.Field("file", "console_test.go").Field("another", "cool story").Error("hello %s", "mars")
+	console.Err(errors.New("one"), "two %s", "three")
+	logger := log.New(log.InfoLevel, console.New(os.Stdout))
+	logger.Err(errors.New("one"), "two %s", "three")
 }
