--- package/log/log_test.go
@@ -1,7 +1,0 @@
-package log_test
-
-import "testing"
-
-func TestTime(t *testing.T) {
-	// defer timer.Debug("TestTime(%s)", t).Error(&err)
-}
