+++ package/log/memory/memory_test.go
@@ -1,20 +1,18 @@
 package memory_test
 
 import (
+	"errors"
 	"testing"
 
 	"github.com/livebud/bud/internal/is"
-	log "github.com/livebud/bud/package/log2"
-	"github.com/livebud/bud/package/log2/memory"
+	log "github.com/livebud/bud/package/log"
+	"github.com/livebud/bud/package/log/memory"
 )
 
 func TestLog(t *testing.T) {
 	is := is.New(t)
 	handler := memory.New()
-	log := &log.Logger{
-		Handler: handler,
-		Level:   log.InfoLevel,
-	}
+	log := log.New(log.InfoLevel, handler)
 	log.Info("hello")
 	log.Info("world")
 	log.Warn("hello %s!", "mars")
@@ -39,4 +37,21 @@ func TestLog(t *testing.T) {
 	is.Equal(fields.Get(fields.Keys()[1]), "memory_test.go")
 	is.Equal(fields.Keys()[2], "one")
 	is.Equal(fields.Get(fields.Keys()[2]), "two")
+}
+
+func TestErr(t *testing.T) {
+	is := is.New(t)
+	handler := memory.New()
+	log := log.New(log.InfoLevel, handler)
+	log.Err(errors.New("one"), "two %s", "three")
+	is.Equal(len(handler.Entries), 1)
+	is.Equal(handler.Entries[0].Level.String(), "error")
+	is.Equal(handler.Entries[0].Message, "two three")
+	fields := handler.Entries[0].Fields
+	is.Equal(len(fields), 2)
+	is.Equal(fields.Keys()[0], "error")
+	is.Equal(fields.Get(fields.Keys()[0]), "one")
+	is.Equal(fields.Keys()[1], "source")
+	source := fields.Get(fields.Keys()[1])
+	is.In(source, "memory_test.TestErr:46") // line number may change
 }
