--- package/log2/filter/filter.go
@@ -1,22 +1,0 @@
-package filter
-
-import (
-	log "github.com/livebud/bud/package/log2"
-)
-
-func New(handler log.Handler, level log.Level) *Handler {
-	return &Handler{handler, level}
-}
-
-// Handler logs by level
-type Handler struct {
-	handler log.Handler
-	level   log.Level
-}
-
-func (f *Handler) Log(entry *log.Entry) error {
-	if entry.Level < f.level {
-		return nil
-	}
-	return f.handler.Log(entry)
-}
